% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getGTExNet}
\alias{getGTExNet}
\title{gteGTExNet - Accessing a network object directly}
\usage{
getGTExNet(tissue)
}
\arguments{
\item{tissue}{One of the labels that can be obtained by calling getGTExTissues() method
to refer to a specific network within the package}
}
\value{
The RDS full file name to the network
}
\description{
gteGTExNet - Accessing a network object directly
}
