% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initDb}
\alias{initDb}
\title{initDb - Initialization of the package so the GTEx networks can be used with
CoExpNets}
\usage{
initDb(mandatory = F)
}
\arguments{
\item{mandatory}{If this parameter is `TRUE` then the networks will be added no matter whether they were already there.}
}
\value{
No value
}
\description{
initDb - Initialization of the package so the GTEx networks can be used with
CoExpNets
}
